requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
haystack-ai==2.0.0
openai==1.14.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
numpy==1.26.4
pydantic==2.6.3 